Rework to interact directly with the Buildkite API
Previously we were just generating YAML on stdout, which the caller was
then expected to upload to Buildkite.

This wasn't adequate to implement the "rollback" behavior, where we need
to also install some metadata onto the build to ensure the deploy will
proceed properly.

So now we interact directly with the agent API, and generate only
informational output for human consumption. We also make use of the
"main" Buildkite API in order to retrieve build metadata for the rollback
functionality.