Add our package dependencies
This repo was unbuildable without knowing the versions of the packages
it was originally coded against. Now that golang has some semi-sane
dependency management we can specify those. Prior to building we can
call 'dep ensure' to get the proper versions in the vendor directory.

This thing should really just be a script that calls out to git and
buildkite-agent, but that's a discussion for the next time we need to
mess with it.