refactor: Upgrade to git2go 34, build statically
Upgrade git2go from 26 to 34. Build it statically with no dependencies so that it does not need a specific version of the libgit2 shared library installed and can run on any version of ubuntu or other linux. Use the goreleaser-cross docker image to build for both linux and MacOS. Upgrade go to 1.20 to match the version of go in the latest goreleaser-cross image. Upgrade the github runner from ubuntu 18.04 to 22.04 since 18.04 has been removed.